+++ packages/icejs/src/require-hook.ts
@@ -28,7 +28,11 @@ export function getHookFiles() {
     'webpack/lib/node/NodeEnvironmentPlugin',
     'webpack/lib/BasicEvaluatedExpression',
     'webpack/lib/ModuleFilenameHelpers',
-    'webpack/lib/GraphHelpers'
+    'webpack/lib/GraphHelpers',
+    'webpack/lib/ExternalsPlugin',
+    'webpack/lib/web/FetchCompileAsyncWasmPlugin',
+    'webpack/lib/web/FetchCompileWasmPlugin',
+    'webpack/lib/ProgressPlugin',
   ];
   const pluginMap = webpackPlugins.map((pluginPath) => {
     const pluginName = getFileName(pluginPath);