+++ packages/swc/src/lib.rs
@@ -10,12 +10,9 @@ use backtrace::Backtrace;
 use napi::{CallContext, Env, JsObject, JsUndefined};
 use std::{env, panic::set_hook, sync::Arc};
 use swc::{Compiler, TransformOutput};
-use swc_common::{
-    self,
-    sync::Lazy,
-    FilePathMapping, SourceMap
-};
+use swc_common::{self, sync::Lazy, FilePathMapping, SourceMap};
 
+mod minify;
 mod transform;
 mod util;
 
@@ -24,7 +21,6 @@ static COMPILER: Lazy<Arc<Compiler>> = Lazy::new(|| {
 
     Arc::new(Compiler::new(cm.clone()))
 });
-
 
 #[module_exports]
 fn init(mut exports: JsObject) -> napi::Result<()> {
@@ -37,6 +33,8 @@ fn init(mut exports: JsObject) -> napi::Result<()> {
 
     exports.create_named_method("transform", transform::transform)?;
     exports.create_named_method("transformSync", transform::transform_sync)?;
+    exports.create_named_method("minify", minify::minify)?;
+    exports.create_named_method("minifySync", minify::minify_sync)?;
 
     Ok(())
 }